create option class
